#152
* remove ticket.called
* Get called ticket user ID and desk id from interactions
