+++ QminderAPI/Classes/Models/Ticket/Created.swift
@@ -1,0 +1,15 @@
+//
+//  Created.swift
+//  QminderAPI
+//
+//  Created by Kristaps Grinbergs on 01/02/2018.
+//
+
+import Foundation
+
+/// Created date
+public struct Created: Codable {
+  
+  /// Time when ticket was created
+  public var date: Date
+}
