+++ QminderAPI/Classes/Models/Ticket/Extra.swift
@@ -1,0 +1,21 @@
+//
+//  Extra.swift
+//  QminderAPI
+//
+//  Created by Kristaps Grinbergs on 01/02/2018.
+//
+
+import Foundation
+
+/// Extra
+public struct Extra: Codable {
+  
+  /// Title
+  public let title: String?
+  
+  /// Value
+  public let value: String?
+  
+  /// URL if there is
+  public let url: String?
+}
