+++ QminderAPI/Classes/Models/Ticket/Served.swift
@@ -1,0 +1,15 @@
+//
+//  Served.swift
+//  QminderAPI
+//
+//  Created by Kristaps Grinbergs on 01/02/2018.
+//
+
+import Foundation
+
+/// Served
+public struct Served: Codable {
+  
+  /// Date of the end of the service
+  public let date: Date?
+}
