+++ fastlane/Fastfile
@@ -2,6 +2,10 @@ update_fastlane
 
 default_platform :ios
 
+before_all do
+  ENV["SLACK_URL"] = "https://hooks.slack.com/services/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"
+end
+
 platform :ios do
   desc "Depoy new Swift API version"
   lane :release do
@@ -18,6 +22,14 @@ platform :ios do
     )
     sh("git fetch --tags")
     pod_push(allow_warnings: true, verbose: true)
+    slack(
+      message: "QminderAPI successfully released!",
+      channel: "#apple-dev-ops",
+      payload: {
+        "Version": version,
+        "Changelog": changelog
+      }
+    )
   end
 
 end
